breakup table into seperate files
