+++ src/app/(post-login)/submission/program/[shortName]/clinical-submission/components/FilesNavigator/types.tsx
@@ -1,0 +1,4 @@
+export type FileRecord = {
+  rowIndex: number;
+  status: "ERROR" | "UPDATE" | "NEW" | "NONE";
+};
